Extracted activate method into a module
- This is needed to be reused in settings module
